User selects one emergency phone number.
The user must select one phone number at a time, instead of a
contact (which may have many phone numbers). As a result, the first
responder has an easier job when dialing the phone number of an
emergency contact (a simple click on the contact). This CL also makes
sure that if the user modifies an emergency contact's number, this is
reflected in the emergency contact preference.

This CL also removes the option of creating a new contact. It makes it
simpler for the user and it has been adviced by UX.
Bug:27461411
Change-Id: I1336aacd2365f51c649ab9be0f505ea9364c4610