Display contact in edit mode
I removed it in the CL that refactored it, but actually it is needed so
that if the phone is unlocked, the first responder (or whoever) can see
the contact and place a call. Otherwise, right now no calls can be
placed without manually looking for each contacts in the contacts app.
Bug:27533454
Change-Id: Ifb7bd120c2f61bca811350dfd89145c1aed7c80e