Little UI changes
These changes include: getting rid of AppCompat delegate stuff,
indenting the title of the activities as suggested by UX, showing a 24dp
avatar for contacts without picture (i.e. do not encircle if they don't
have a picture), and set the width of the tabs to be different in landscape
and portrait mode.
Bug:27533454
Change-Id: I969cd267845bb061803688a047b87407814fcee8