Allow the EmergencyInfo app to be tweaked with an overlay.
This allows the EmergencyInfo app to be replaced on some devices with
another app with the same package name. The two apps will share much of
the same source. The FeatureFactory interface provides a form of
dependency injection, allowing distinct concrete implementations to
exist for the same interface in the different apps.

In this first commit, the fragment class for the Contacts tab can be
overridden. This is an interim state intended to demonstrate the use of
the overlay mechanism in this app, while causing no regressions.

Test: Ran existing unit tests. Manually verified app still works as
expected.

Bug: 35335027
Change-Id: Icad03cf3351e47378c9b6a4b304621546d4e40c2