Convert NameAutoCompletePreferenceTest to a unit test.
The instrumentation test runner is deprecated.
Rewriting the tests paves the way to migrating to the support library,
which the new UX relies on.

This also removes the test case that verifies that a preference dialog is
no longer shown after clicking either button, because the support library
version of PreferenceDialog gives us no means to test this. Further, the
dismissal behavior is a property of the base class, so removing this
test case has no impact on test coverage.

Bug: 30046624
Test: runtest --path packages/apps/EmergencyInfo/tests
Change-Id: I263d35a56de7d9034d86e5cd4510672286287521