Show text changes immediately.
In the support library version of Preference, settings changes aren't
visible immediately. Setting the summary value directly works around the
bug.

There's one bug this doesn't fix: when editing a value from set
(non-empty) to the empty string, the UX displays the empty string rather
than the default summary value. Since the default summary value is set
in the XML file, and setSummary clobbers it, there's no way to get
access to it. I felt this additional bug is minor enough, and the
changes required to fix it substantial enough, that it's not worth
fixing in this release.

Change-Id: Ife521f5430476a211d8eb4834c87d5a4a22d0025
Fix: 63386208
Test: runtest --path packages/apps/EmergencyInfo/tests/