Hide Emergency Info summary values from search.
Because the summary value displayed by the fragment is dynamically
updated, while the settings search index is statically computed from the
preference values, the two cannot presently be in sync. By renaming the
summary preference attribute, the search index omits the summary, which
prevents it from displaying an incorrect value.

Change-Id: I6598d92ea34e55bcc885050888faa8123233987b
Bug: 64654734
Fix: 64654734
Test: manual, and runtest --path packages/apps/EmergencyInfo/