EmergencyInfo: Remove direct dep on sdk_vcurrent module.
This particular module is going away soon. If a dependency on the SDK
is needed, the right way to specify that is via the LOCAL_SDK_VERSION
variable.

Also remove the @SmallTest annotation which were imported from this
jar file. They don't appear to be needed for the tests to run.

This change is a partial CP of I8aeab1016dd51e2985add5fbec89edd70d2cc0a3.

Test: make RunEmergencyInfoRoboTests
Bug: 77525052
Change-Id: I55128ad80d974a3f95588e46b919a8684683fb79
Merged-In: I8aeab1016dd51e2985add5fbec89edd70d2cc0a3