Fix Talkback speak n+1 items on medical info view
The design was an 8dp padding between ActionBar and PreferenceList. The
original implementation was adding an empty Preference with 8dp hight to
have the padding effect, but Talkback would take the item into account.
So the Talkback would always speak n+1 items.

Solution: add a top padding (8dp) to the Preference View, and remove
the empty preference.

Bug: 118583643
Test: Manually
Change-Id: I159309e8a5c5e87f4fe0d8ec48e54c88f79848e9